+++ packages/core/inline-dialog/src/InlineDialog/__tests__/unit/server-side-hydrate.tsx
@@ -31,8 +31,8 @@ jest.mock('popper.js', () => {
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate inline-dialog correctly', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate inline-dialog correctly', async () => {
   const [example] = await getExamplesFor('inline-dialog');
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require