+++ packages/core/pagination/src/__tests__/unit/server-side-hydrate.tsx
@@ -15,10 +15,10 @@ beforeEach(() => {
 afterEach(() => {
   jest.resetAllMocks();
 });
-
-test('should ssr then hydrate pagination correctly', async () => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate pagination correctly', async () => {
   const [example] = await getExamplesFor('pagination');
-  // $StringLitteral
+
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');