+++ packages/media/media-core/src/__tests__/unit/server-side-hydrate.tsx
@@ -14,8 +14,8 @@ beforeEach(() => {
   jest.setTimeout(20000);
 });
 
-// TODO: failing due to whatwg-fetch issues. Re-enable once fixed
-test('should ssr then hydrate media-core correctly', async () => {
+// TODO: failing due to whatwg-fetch issues. Re-enable once fixed// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('should ssr then hydrate media-core correctly', async () => {
   const [example] = await getExamplesFor('media-core');
   const Example = await require(example.filePath).default;