+++ packages/search/global-search/src/__tests__/unit/mocks/_mockPrefetchResults.ts
@@ -1,4 +1,8 @@
 import { ConfluenceRecentsMap, Result } from '../../../model/Result';
+import {
+  SearchResultsMap,
+  EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE,
+} from '../../../api/CrossProductSearchClient';
 
 const confluenceRecentItemsPromise: Promise<
   ConfluenceRecentsMap
@@ -16,11 +20,33 @@ const confluenceRecentItemsPromise: Promise<
     totalSize: 0,
   },
 });
+
+const crossProductRecentItemsPromise: Promise<
+  SearchResultsMap
+> = Promise.resolve({
+  ...EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE.results,
+  'jira.issue': {
+    items: [],
+    totalSize: 0,
+  },
+  'jira.board,project,filter': {
+    items: [],
+    totalSize: 0,
+  },
+} as SearchResultsMap);
+
 const abTestPromise: Promise<Result[]> = Promise.resolve([]);
 
 export const mockConfluencePrefetchedData = jest.fn(() => {
   return {
     confluenceRecentItemsPromise,
+    abTestPromise,
+  };
+});
+
+export const mockJiraPrefetchedData = jest.fn(() => {
+  return {
+    crossProductRecentItemsPromise,
     abTestPromise,
   };
 });