+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/global-navigation
 
+## 7.4.0
+
+### Minor Changes
+
+- [minor][9fe5a0fa58](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9fe5a0fa58):
+
+  Allow app to configure a backIcon for drawer component inside global navigation
+
 ## 7.3.9
 
 ### Patch Changes